requirements.txt
python-dotenv
openai
langchain
gradio
